backpressed error fix
